+++ pkg/processor/sumologicschemaprocessor/config.go
@@ -22,19 +22,39 @@ import (
 type Config struct {
 	config.ProcessorSettings `mapstructure:",squash"`
 
-	AddCloudNamespace           bool `mapstructure:"add_cloud_namespace"`
-	TranslateAttributes         bool `mapstructure:"translate_attributes"`
-	TranslateTelegrafAttributes bool `mapstructure:"translate_telegraf_attributes"`
+	AddCloudNamespace           bool                    `mapstructure:"add_cloud_namespace"`
+	TranslateAttributes         bool                    `mapstructure:"translate_attributes"`
+	TranslateTelegrafAttributes bool                    `mapstructure:"translate_telegraf_attributes"`
+	NestAttributes              *NestingProcessorConfig `mapstructure:"nest_attributes"`
+	AggregateAttributes         []aggregationPair       `mapstructure:"aggregate_attributes"`
+}
+
+type aggregationPair struct {
+	Attribute string   `mapstructure:"attribute"`
+	Patterns  []string `mapstructure:"prefixes"`
 }
 
 const (
 	defaultAddCloudNamespace           = true
 	defaultTranslateAttributes         = true
 	defaultTranslateTelegrafAttributes = true
+
+	// Nesting processor default config
+	defaultNestingEnabled            = false
+	defaultNestingSeparator          = "."
+	defaultNestingSquashSingleValues = false
+)
+
+var (
+	defaultAggregateAttributes = []aggregationPair{}
 )
 
 // Ensure the Config struct satisfies the config.Processor interface.
-var _ component.Config = (*Config)(nil)
+var (
+	_                     component.Config = (*Config)(nil)
+	defaultNestingInclude                  = []string{}
+	defaultNestingExclude                  = []string{}
+)
 
 func createDefaultConfig() component.Config {
 	return &Config{
@@ -42,6 +62,14 @@ func createDefaultConfig() component.Config {
 		AddCloudNamespace:           defaultAddCloudNamespace,
 		TranslateAttributes:         defaultTranslateAttributes,
 		TranslateTelegrafAttributes: defaultTranslateTelegrafAttributes,
+		NestAttributes: &NestingProcessorConfig{
+			Separator:          defaultNestingSeparator,
+			Enabled:            defaultNestingEnabled,
+			Include:            defaultNestingInclude,
+			Exclude:            defaultNestingExclude,
+			SquashSingleValues: defaultNestingSquashSingleValues,
+		},
+		AggregateAttributes: defaultAggregateAttributes,
 	}
 }