+++ CHANGELOG.md
@@ -7,7 +7,29 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-[Unreleased]: https://github.com/SumoLogic/sumologic-otel-collector/compare/v0.66.0-sumo-0...main
+### Added
+
+- feat(sumologicschemaprocessor): add nesting processor [#877]
+- feat(sumologicschemaprocessor): add allowlist and denylist to nesting processor [#880]
+- feat(sumologicschemaprocessor) allow aggregating attributes with given name patterns [#871]
+- feat(sumologicschemaprocessor): add squashing single values in nesting processor [#881]
+- feat: update collector metadata on start [#858]
+
+[Unreleased]: https://github.com/SumoLogic/sumologic-otel-collector/compare/v0.67.0-sumo-0...main
+[#858]: https://github.com/SumoLogic/sumologic-otel-collector/pull/858
+[#877]: https://github.com/SumoLogic/sumologic-otel-collector/pull/877
+[#880]: https://github.com/SumoLogic/sumologic-otel-collector/pull/880
+[#871]: https://github.com/SumoLogic/sumologic-otel-collector/pull/871
+[#881]: https://github.com/SumoLogic/sumologic-otel-collector/pull/881
+
+## [v0.67.0-sumo-0]
+
+### Released 2022-12-19
+
+### Added
+
+- feature(packaging/msi): add conf.d dir, mv token/tags to common.yaml [869]
+- feat(ci): build msi packages for dev & release jobs [#856]
 
 ### Added
 
@@ -21,9 +43,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - fix(otelcolbuilder): use correct upstream modules [#864]
 
-[#858]: https://github.com/SumoLogic/sumologic-otel-collector/pull/858
+[v0.67.0-sumo-0]: https://github.com/SumoLogic/sumologic-otel-collector/compare/v0.66.0-sumo-0...v0.67.0-sumo-0
 [#864]: https://github.com/SumoLogic/sumologic-otel-collector/pull/864
 [#867]: https://github.com/SumoLogic/sumologic-otel-collector/pull/867
+[#869]: https://github.com/SumoLogic/sumologic-otel-collector/pull/869
+[#856]: https://github.com/SumoLogic/sumologic-otel-collector/pull/856
 
 ## [v0.66.0-sumo-0]